#396 Updated anvil-catalog to follow the configuration guide, except with using the anvil response model
