+++ explorer/app/views/Index/index.tsx
@@ -78,7 +78,7 @@ export const Index = (props: AzulEntitiesStaticResponse): JSX.Element => {
   const [tabsValue, setTabsValue] = useState<TabsValue>(route);
   const tabs = getTabs(entities);
 
-  // Fetch summary and entities.
+  // Fetch summary and workspaceEntityConfig.
   const { response: summaryResponse } = useSummary();
   const { categories, loading, pagination, response, setFilter, sort } =
     useFetchEntities(props, []);
