+++ spa/src/app/files/_ngrx/files.selectors.ts
@@ -30,7 +30,7 @@ export const selectFileSummary =
 
 /**
  * Returns complete table state, including state that has been cached for the current selected entity, and any previously
- * selected entities.
+ * selected workspaceEntityConfig.
  *
  * @returns {TableState}
  */
