+++ spa/src/app/files/_ngrx/search/search.reducer.ts
@@ -2,7 +2,7 @@
  * Human Cell Atlas
  * https://www.humancellatlas.org/
  *
- * Search reducer, handles actions related to selecting and deselecting file facet terms and entities.
+ * Search reducer, handles actions related to selecting and deselecting file facet terms and workspaceEntityConfig.
  */
 
 // Core dependencies
